Addressed PR Comments
