+++ backend/django_app/settings.py
@@ -16,6 +16,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "product",
+    # "product.apps.ProductConfig",
     "rest_framework",
     'django_filters'
 ]
